+++ packages/ui/src/index.ts
@@ -3,3 +3,4 @@ export * from "./components/Button";
 export * from "./components/Card";
 export * from "./components/Field";
 export * from "./components/StatusBadge";
+export * from "./lib/utils";